refactor: SortList 컴포넌트 공용화
