+++ frontend/src/domains/sort/article.ts
@@ -1,0 +1,4 @@
+export const ARTICLE_SORT_MAP = {
+  createdAt: "최신순",
+  views: "조회순",
+} as const;
