+++ frontend/src/domains/sort/project.ts
@@ -1,0 +1,5 @@
+export const PROJECT_SORT_MAP = {
+  createdAt: "최신순",
+  views: "조회순",
+  loves: "좋아요순",
+} as const;
